Research Data/Experimental/Experimental-3e970bfea0fb85405b07cd96dfc6e1d1ca862ed2/requirements.txt
atomic-agents
chromadb
crewai
crewai_tools
docling
easyocr
Flask==3.0.0
instructor
langchain
langchain-chroma
langchain-classic
langchain-community
langchain_openai
langchain-text-splitters
openai
pillow
pypdf
pymongo
python-dotenv==1.0.0
pytest
pytest-asyncio
pytest-mock
scikit-learn
streamlit
torch
transformers
voyageai